rearranged all dynamic calibration functions into new dynamic_spinup module, added tests for dynamic mu star calibration
