+++ oggm/params.cfg
@@ -170,8 +170,8 @@ smooth_widths_window_size = 1
 # Leave empty if you want to do your own cuisine.
 baseline_climate = CRU
 # Hydrological year definition
-hydro_month_nh = 1
-hydro_month_sh = 1
+hydro_month_nh = 10
+hydro_month_sh = 4
 # specify here the start and end year where oggm will search for tstar
 # candidates (note that the window will be reduced by mu_star_halfperiod on
 # each side of the window). Set to 0, 0 for the default (the entire available
@@ -203,14 +203,14 @@ temp_all_solid = 0.
 temp_all_liq = 2.
 temp_melt = -1.
 # precipitation correction: set to a float for a constant scaling factor
-prcp_scaling_factor = 1.6
+prcp_scaling_factor = 2.5
 # Should we use the default, pre-calibrated reference tstars or are we
 # running the calibration ourselves? The default should be False, which
 # raises a warning when trying to calibrate.
 run_mb_calibration = False
 # historical climate quality check - this parameter ensures that there is
 # at least N months per year where melt and/or accumulation can occur
-climate_qc_months = 0
+climate_qc_months = 3
 # Bounds on mu*
 # Values out of these limits are considered bad and will lead to an error
 # When matching geodetic MB on a glacier per glacier basis, this should
