+++ oggm/tasks.py
@@ -57,8 +57,8 @@ from oggm.core.flowline import run_random_climate
 from oggm.core.flowline import run_from_climate_data
 from oggm.core.flowline import run_constant_climate
 from oggm.core.flowline import run_with_hydro
-from oggm.core.flowline import run_dynamic_spinup
-from oggm.core.flowline import dynamic_mu_star_calibration
+from oggm.core.dynamic_spinup import run_dynamic_spinup
+from oggm.core.dynamic_spinup import run_dynamic_mu_star_calibration
 from oggm.utils import copy_to_basedir
 from oggm.utils import gdir_to_tar
 from oggm.utils import merge_consecutive_run_outputs
